Support setting a per-color attachment format
Users can now specify a different format for each color attachment of a
buffer. Additionally, we simplify the buffer config schema, by
eliminating the attachment count and format keys, and roll it all into
the buffer key. This commit should be sort-of backwards compatible. Any
config that specified buffer::format will need to be updated, but that's
OK as I think we should get a parse errror/warning in those cases.